Make it easier to write transforms
