+++ transform-function/src/main/resources/config.yml
@@ -1,7 +1,7 @@
 tenant: public
 namespace: default
-name: user-transform-new-13
-className: com.bridge.data.TransformStringMessage
+name: user-transform-new-14
+className: com.bridge.data.TransformUser
 inputs:
   - "persistent://public/default/debezium.public.users"
 output: "persistent://public/default/debezium.public.users-public"
